Refactor code & optimise folder structure
