+++ src/fastFingersGameApp.scss
